move CLI utility functions to CLI tests
